refactor: cleanup deprecated-related APIs in `std/io` and `std/streams`
