+++ io/string_writer.ts
@@ -15,7 +15,7 @@ const decoder = new TextDecoder();
  *   StringReader,
  *   StringWriter,
  * } from "https://deno.land/std@$STD_VERSION/io/mod.ts";
- * import { copy } from "https://deno.land/std@$STD_VERSION/streams/copy.ts";
+ * import { copy } from "https://deno.land/std@$STD_VERSION/io/copy.ts";
  *
  * const w = new StringWriter("base");
  * const r = new StringReader("0123456789");
