+++ streams/writable_stream_from_writer.ts
@@ -1,20 +1,13 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import { writeAll } from "./write_all.ts";
-import type { Closer, Writer } from "../io/types.ts";
-
-function isCloser(value: unknown): value is Closer {
-  return typeof value === "object" && value !== null && value !== undefined &&
-    "close" in value &&
-    // deno-lint-ignore no-explicit-any
-    typeof (value as Record<string, any>)["close"] === "function";
-}
+import type { Writer } from "../io/types.ts";
+import { toWritableStream } from "../io/to_writable_stream.ts";
 
 /**
  * Options for {@linkcode writableStreamFromWriter}.
  *
- * @deprecated (will be removed after 1.0.0) Use {@linkcode WritableStream} directly.
+ * @deprecated (will be removed after 1.0.0) Use {@linkcode toWritableStream} instead.
  */
 export interface WritableStreamFromWriterOptions {
   /**
@@ -29,34 +22,11 @@ export interface WritableStreamFromWriterOptions {
 /**
  * Create a {@linkcode WritableStream} from a {@linkcode Writer}.
  *
- * @deprecated (will be removed after 1.0.0) Use {@linkcode WritableStream} directly.
+ * @deprecated (will be removed after 1.0.0) Use {@linkcode toWritableStream} instead.
  */
 export function writableStreamFromWriter(
   writer: Writer,
   options: WritableStreamFromWriterOptions = {},
 ): WritableStream<Uint8Array> {
-  const { autoClose = true } = options;
-
-  return new WritableStream({
-    async write(chunk, controller) {
-      try {
-        await writeAll(writer, chunk);
-      } catch (e) {
-        controller.error(e);
-        if (isCloser(writer) && autoClose) {
-          writer.close();
-        }
-      }
-    },
-    close() {
-      if (isCloser(writer) && autoClose) {
-        writer.close();
-      }
-    },
-    abort() {
-      if (isCloser(writer) && autoClose) {
-        writer.close();
-      }
-    },
-  });
+  return toWritableStream(writer, options);
 }
