+++ streams/write_all.ts
@@ -1,7 +1,10 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import { writeAll as _writeAll } from "../io/write_all.ts";
+import {
+  writeAll as _writeAll,
+  writeAllSync as _writeAllSync,
+} from "../io/write_all.ts";
 import type { Writer, WriterSync } from "../io/types.ts";
 export type { Writer, WriterSync };
 
@@ -62,8 +65,5 @@ export async function writeAll(w: Writer, arr: Uint8Array) {
  * @deprecated (will be removed in 0.214.0) Import from {@link https://deno.land/std/io/write_all.ts} instead.
  */
 export function writeAllSync(w: WriterSync, arr: Uint8Array) {
-  let nwritten = 0;
-  while (nwritten < arr.length) {
-    nwritten += w.writeSync(arr.subarray(nwritten));
-  }
+  _writeAllSync(w, arr);
 }
